Update matplotlib_assignment.py
Firstly importing the libraries
1.
Defining x and y and using plt.plot() to plot a line chart and furthur defining the title and lables.

2.
Defining students and marks and using plt.bar() to plot a line chart and furthur defining the title and lables.

3.
Defining regions and revenue and using plt.pie() with autopct depicts the percentage displayed with decimal and explode depicts which part should be taken and furthur defining the title and lables.

4.
Defining no. from 1 to 101 using random.randint() function and ploting them using plt.hist() and furthur defining the title and lables.